% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.map.R
\name{read.map}
\alias{read.map}
\title{Read 'MAP' vector polylist}
\usage{
read.map(filename)
}
\arguments{
\item{filename}{}
}
\value{
polylist
}
\description{
See https://stackoverflow.com/questions/61614314/how-to-read-a-map-file-extension-in-r
}
\examples{
## WIP doesn't work generally needs finish
x <- read.map(system.file("extdata/se_municip.MAP", package = "polylist", mustWork = TRUE))
df <- setNames(as.data.frame(do.call(rbind, x)), c("x", "y"))
df$region.name <- rep(attr(x, "region.name"), unlist(lapply(x, nrow)))
## in case there are multi-rings
df$linestring_id <- cumsum(c(0, diff(is.na(df$x))))
df$polygon_id <- as.integer(factor(df$region.name))
df <- df[!is.na(df$x), ]
sfx <- sfheaders::sf_polygon(df, x = "x", y = "y", "linestring_id" = "linestring_id",
  polygon_id = "polygon_id", keep = TRUE)
#sf::st_crs(sfx) <- sf::st_crs(<whatever it is probably 4326>)
}
